Phase 2.2: split fetchFunctions.ts into utils/api/ domain modules
The 408-line fetchFunctions.ts becomes utils/api/ with the base fetch
helpers + createEntityClient in client.ts and one module per domain
(resources, topics, providers, domains, radar, routines, dailies, tasks,
tags, modules, interactions, templates, admin). fetchFunctions.ts stays
as a one-line shim so no importers change.
