+++ packages/client/src/utils/api/interactions.ts
@@ -1,0 +1,13 @@
+import type { Interaction } from "@emstack/types";
+
+import { createEntityClient } from "./client";
+
+export const interactionsApi = createEntityClient<Interaction>(
+  "interactions",
+  "interaction",
+);
+
+export const fetchInteractions = interactionsApi.list;
+export const upsertInteraction = interactionsApi.upsert;
+export const createInteraction = interactionsApi.create;
+export const deleteSingleInteraction = interactionsApi.delete;
