+++ packages/client/src/utils/fetchFunctions.ts
@@ -1,408 +1,2 @@
-import type {
-  Resource,
-  ResourceInResources,
-  TopicForTopicsPage,
-  CourseProvider,
-  DailyCriteriaTemplate,
-  Domain,
-  Daily,
-  Interaction,
-  Module,
-  ModuleGroup,
-  Radar,
-  Routine,
-  RoutineTemplate,
-  Tag,
-  TagGroup,
-  Task,
-  TaskType,
-  OnboardData,
-  Topic,
-} from "@emstack/types";
-
-interface SuccessObj {
-  status: string;
-}
-
-interface CreateResponse {
-  status: string;
-  id: string;
-}
-
-async function fetchJson<T>(url: string): Promise<T> {
-  const response = await fetch(url);
-  if (!response.ok) {
-    throw new Error(
-      `Request to ${url} failed (${response.status} ${response.statusText})`,
-    );
-  }
-  return await response.json();
-}
-
-async function postJson<T>(
-  url: string,
-  data?: object,
-  errorLabel?: string,
-): Promise<T> {
-  const response = await fetch(url, {
-    method: "POST",
-    ...(data === undefined
-      ? {}
-      : {
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(data),
-      }),
-  });
-  if (!response.ok) {
-    const body = await response.text().catch(() => "");
-    throw new Error(
-      `${errorLabel ?? `POST ${url}`} failed (${response.status} ${response.statusText})${body ? `: ${body}` : ""}`,
-    );
-  }
-  return await response.json();
-}
-
-async function putJson<T>(
-  url: string,
-  data: object,
-  errorLabel?: string,
-): Promise<T> {
-  const response = await fetch(url, {
-    method: "PUT",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(data),
-  });
-  if (!response.ok) {
-    const body = await response.text().catch(() => "");
-    throw new Error(
-      `${errorLabel ?? `PUT ${url}`} failed (${response.status} ${response.statusText})${body ? `: ${body}` : ""}`,
-    );
-  }
-  return await response.json();
-}
-
-async function deleteJson<T>(url: string, errorLabel?: string): Promise<T> {
-  const response = await fetch(url, {
-    method: "DELETE",
-  });
-  if (!response.ok) {
-    const text = await response.text().catch(() => "");
-    let serverMessage: string | undefined;
-    if (text) {
-      try {
-        const parsed = JSON.parse(text);
-        if (parsed && typeof parsed.message === "string") {
-          serverMessage = parsed.message;
-        }
-      }
-      catch {
-        serverMessage = text;
-      }
-    }
-    throw new Error(
-      serverMessage
-      ?? `${errorLabel ?? `DELETE ${url}`} failed (${response.status} ${response.statusText})`,
-    );
-  }
-  return await response.json();
-}
-
-interface EntityClient<TEntity, TList> {
-  list: () => Promise<TList>;
-  get: (id: string) => Promise<TEntity>;
-  create: (data: Record<string, unknown>) => Promise<CreateResponse>;
-  upsert: (id: string, data: Record<string, unknown>) => Promise<SuccessObj>;
-  delete: (id: string) => Promise<SuccessObj>;
-  duplicate: (id: string) => Promise<CreateResponse>;
-}
-
-function createEntityClient<TEntity, TList = TEntity[]>(
-  endpoint: string,
-  label: string,
-): EntityClient<TEntity, TList> {
-  const base = `/api/${endpoint}`;
-  return {
-    list: () => fetchJson<TList>(base),
-    get: id => fetchJson<TEntity>(`${base}/${id}`),
-    create: data =>
-      postJson<CreateResponse>(base, data, `Failed to create ${label}`),
-    upsert: (id, data) =>
-      putJson<SuccessObj>(`${base}/${id}`, data, `Failed to update ${label}`),
-    delete: id =>
-      deleteJson<SuccessObj>(`${base}/${id}`, `Failed to delete ${label}`),
-    duplicate: id =>
-      postJson<CreateResponse>(
-        `${base}/${id}/duplicate`,
-        undefined,
-        `Failed to duplicate ${label}`,
-      ),
-  };
-}
-
-export const resourcesApi = createEntityClient<Resource, ResourceInResources[]>(
-  "resources",
-  "resource",
-);
-export const topicsApi = createEntityClient<Topic, TopicForTopicsPage[]>(
-  "topics",
-  "topic",
-);
-export const providersApi = createEntityClient<CourseProvider>(
-  "providers",
-  "provider",
-);
-export const domainsApi = createEntityClient<Domain>("domains", "domain");
-export const routinesApi = createEntityClient<Routine>("routines", "routine");
-export const tasksApi = createEntityClient<Task>("tasks", "task");
-export const taskTypesApi = createEntityClient<TaskType>(
-  "task-types",
-  "task type",
-);
-export const tagGroupsApi = createEntityClient<TagGroup>(
-  "tag-groups",
-  "tag group",
-);
-export const tagsApi = createEntityClient<Tag>("tags", "tag");
-export const moduleGroupsApi = createEntityClient<ModuleGroup>(
-  "module-groups",
-  "module group",
-);
-export const modulesApi = createEntityClient<Module>("modules", "module");
-export const interactionsApi = createEntityClient<Interaction>(
-  "interactions",
-  "interaction",
-);
-export const dailyCriteriaTemplatesApi = createEntityClient<DailyCriteriaTemplate>(
-  "daily-criteria-templates",
-  "criteria template",
-);
-export const routineTemplatesApi = createEntityClient<RoutineTemplate>(
-  "routine-templates",
-  "routine template",
-);
-
-export const fetchTopics = topicsApi.list;
-export const fetchProviders = providersApi.list;
-export const fetchResources = resourcesApi.list;
-export const fetchDomains = domainsApi.list;
-// Dailies are now routines projected into the Daily shape. `projected=true`
-// returns BOTH daily- and weekly-mode routines (each carrying its `mode`) so the
-// tracker / dashboard list weekly routines too. These thin wrappers keep the
-// existing daily tracker / dashboard / detail components working unchanged.
-export const fetchDailies = () =>
-  fetchJson<Daily[]>("/api/routines?projected=true");
-export const fetchRoutines = routinesApi.list;
-export const fetchTasks = tasksApi.list;
-export const fetchTaskTypes = taskTypesApi.list;
-export const fetchTagGroups = tagGroupsApi.list;
-export const fetchTags = tagsApi.list;
-export const fetchModuleGroups = moduleGroupsApi.list;
-export const fetchModules = modulesApi.list;
-export const fetchInteractions = interactionsApi.list;
-export const fetchDailyCriteriaTemplates = dailyCriteriaTemplatesApi.list;
-export const fetchRoutineTemplates = routineTemplatesApi.list;
-
-export const fetchSingleResource = resourcesApi.get;
-export const fetchSingleTopic = topicsApi.get;
-export const fetchSingleProvider = providersApi.get;
-export const fetchSingleDomain = domainsApi.get;
-export const fetchSingleDaily = (id: string) =>
-  fetchJson<Daily>(`/api/routines/${id}`);
-export const fetchSingleRoutine = routinesApi.get;
-export const fetchSingleTask = tasksApi.get;
-
-export const upsertResource = resourcesApi.upsert;
-export const upsertTopic = topicsApi.upsert;
-export const upsertProvider = providersApi.upsert;
-export const upsertDomain = domainsApi.upsert;
-// No mode injection: the routines upsert merges partially, so an absent `mode`
-// preserves the row's existing mode. This keeps completion edits safe on both
-// daily- and weekly-mode routines (a weekly routine is never flipped to daily).
-export const upsertDaily = routinesApi.upsert;
-export const upsertRoutine = routinesApi.upsert;
-export const upsertTask = tasksApi.upsert;
-export const upsertTaskType = taskTypesApi.upsert;
-export const upsertTagGroup = tagGroupsApi.upsert;
-export const upsertTag = tagsApi.upsert;
-export const upsertModuleGroup = moduleGroupsApi.upsert;
-export const upsertModule = modulesApi.upsert;
-export const upsertInteraction = interactionsApi.upsert;
-export const upsertDailyCriteriaTemplate = dailyCriteriaTemplatesApi.upsert;
-export const upsertRoutineTemplate = routineTemplatesApi.upsert;
-
-export const createTopic = topicsApi.create;
-export const createProvider = providersApi.create;
-export const createDomain = domainsApi.create;
-export const createRoutine = routinesApi.create;
-export const createTask = tasksApi.create;
-export const createTaskType = taskTypesApi.create;
-export const createTagGroup = tagGroupsApi.create;
-export const createTag = tagsApi.create;
-export const createModuleGroup = moduleGroupsApi.create;
-export const createModule = modulesApi.create;
-export const createInteraction = interactionsApi.create;
-export const createDailyCriteriaTemplate = dailyCriteriaTemplatesApi.create;
-export const createRoutineTemplate = routineTemplatesApi.create;
-
-export const deleteSingleResource = resourcesApi.delete;
-export const deleteSingleTopic = topicsApi.delete;
-
-export async function bulkDeleteTopics(
-  ids: string[],
-): Promise<{ status: string;
-  count: number; }> {
-  return postJson(
-    "/api/topics/bulk-delete",
-    {
-      ids,
-    },
-    "Failed to delete topics",
-  );
-}
-
-export const deleteSinglePlatform = providersApi.delete;
-export const deleteSingleDomain = domainsApi.delete;
-export const deleteSingleDaily = routinesApi.delete;
-export const deleteSingleRoutine = routinesApi.delete;
-export const deleteSingleTask = tasksApi.delete;
-export const deleteSingleTaskType = taskTypesApi.delete;
-export const deleteSingleTagGroup = tagGroupsApi.delete;
-export const deleteSingleTag = tagsApi.delete;
-export const deleteSingleModuleGroup = moduleGroupsApi.delete;
-export const deleteSingleModule = modulesApi.delete;
-export const deleteSingleInteraction = interactionsApi.delete;
-export const deleteSingleDailyCriteriaTemplate = dailyCriteriaTemplatesApi.delete;
-export const deleteSingleRoutineTemplate = routineTemplatesApi.delete;
-
-export const duplicateResource = resourcesApi.duplicate;
-export const duplicateDomain = domainsApi.duplicate;
-export const duplicateRoutine = routinesApi.duplicate;
-
-export async function fetchSeed(): Promise<SuccessObj> {
-  return await fetch("/api/seed").then(res => res.json());
-}
-
-export async function fetchClear(): Promise<SuccessObj> {
-  return await fetch("/api/clearData").then(res => res.json());
-}
-
-export async function incrementResourceProgress(
-  id: string,
-): Promise<{
-  status: string;
-  id: string;
-  progressCurrent: number;
-  progressTotal: number;
-}> {
-  return postJson(
-    `/api/resources/${id}/incrementProgress`,
-    undefined,
-    "Failed to increment resource progress",
-  );
-}
-
-export async function postOnboardForm(
-  formData: OnboardData,
-): Promise<SuccessObj> {
-  return await fetch("/api/submitOnboardData", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(formData),
-  }).then(res => res.json());
-}
-
-export async function fetchRadar(domainId: string): Promise<Radar> {
-  return fetchJson<Radar>(`/api/domains/${domainId}/radar`);
-}
-
-interface RadarConfigPayload {
-  quadrants: { id?: string;
-    name: string;
-    position: number; }[];
-  rings: { id?: string;
-    name: string;
-    position: number;
-    isAdopted?: boolean; }[];
-  hasAdoptedSection?: boolean;
-}
-
-export async function upsertRadarConfig(
-  domainId: string,
-  data: RadarConfigPayload,
-): Promise<SuccessObj> {
-  return putJson(
-    `/api/domains/${domainId}/radar`,
-    data,
-    "Failed to save radar config",
-  );
-}
-
-interface BlipPayload {
-  topicId: string;
-  description?: string | null;
-  quadrantId?: string | null;
-  ringId?: string | null;
-  isIgnored?: boolean | null;
-}
-
-export async function createRadarBlip(
-  domainId: string,
-  data: BlipPayload,
-): Promise<CreateResponse> {
-  return postJson(
-    `/api/domains/${domainId}/radar/blips`,
-    data,
-    "Failed to create blip",
-  );
-}
-
-export async function upsertRadarBlip(
-  domainId: string,
-  blipId: string,
-  data: BlipPayload,
-): Promise<SuccessObj> {
-  return putJson(
-    `/api/domains/${domainId}/radar/blips/${blipId}`,
-    data,
-    "Failed to update blip",
-  );
-}
-
-export async function deleteRadarBlip(
-  domainId: string,
-  blipId: string,
-): Promise<SuccessObj> {
-  return deleteJson(
-    `/api/domains/${domainId}/radar/blips/${blipId}`,
-    "Failed to delete blip",
-  );
-}
-
-export interface BulkBlipEntry {
-  topicId?: string | null;
-  newTopicName?: string | null;
-  newTopicDescription?: string | null;
-  description?: string | null;
-  quadrantId?: string | null;
-  ringId?: string | null;
-}
-
-export async function bulkCreateRadarBlips(
-  domainId: string,
-  data: { blips: BulkBlipEntry[] },
-): Promise<{ status: string;
-  count: number;
-  ids: string[];
-  skippedDuplicates?: number; }> {
-  return postJson(
-    `/api/domains/${domainId}/radar/blips/bulk`,
-    data,
-    "Failed to bulk-create blips",
-  );
-}
+// Kept as a shim so existing imports keep working; the clients live in ./api.
+export * from "./api";
